Change datatype of fromMessageId to long from int

+++ src/main/java/dev/manav/telefetch/model/ChatHistoryTemplate.java
@@ -7,6 +7,6 @@ import lombok.Setter;
 @Setter
 public class ChatHistoryTemplate {
     private long chatId;
-    private int fromMessageId;
+    private long fromMessageId;
     private int limit;
 }
